dao와 dto 수정 및 업데이트

+++ rest/src/main/java/com/jamjam/rest/dto/User.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class User {
 	
-	private Long user_id;
+	private Integer user_id;
 	private String email;
 	private String name;
 	private String contact;
